Impl async framework

+++ src/arch/hosted/time.rs
@@ -1,3 +1,3 @@
-pub fn get_time() -> usize {
-    std::time::SystemTime::now().duration_since(std::time::UNIX_EPOCH).unwrap().as_nanos() as usize
+pub fn get_time() -> u64 {
+    std::time::SystemTime::now().duration_since(std::time::UNIX_EPOCH).unwrap().as_nanos() as u64
 }
